Update response objects and headers. Add error checking for not-existent users for get_user_info(), delete_user(), and update_user().

+++ app/users/routes.py
@@ -5,7 +5,7 @@ from flask_sqlalchemy import SQLAlchemy
 from werkzeug.security import generate_password_hash, check_password_hash
 from ..models import User
 from datetime import datetime
-# from flask_login import login_required
+from flask_login import login_required
 from .. import db
 
 
@@ -35,7 +35,7 @@ def add_user():
             response = {
                 "message": f"{username} ({email}) already exists!",
                 "success": False
-                        }
+            }
             return jsonify(response), 409, headers
 
         new_user = User(first_name=first_name,
@@ -65,46 +65,70 @@ def get_user_info():
     # get user info
     user = User.query.get(user_id)
 
-    # construct response
-    response = []
-    response.append({
-        "user_id": user.user_id,                
-        "user_first_name": user.first_name,
-        "user_last_name": user.last_name,
-        "user_username": user.username,
-        "user_email": user.email,
-        "user_date_created": user.date_created
-    })
+    # prepare headers for response
+    headers = {"Content-Type": "application/json"}
 
-    return jsonify(response)
+    # get user info if user_id exists in the database
+    if user:
+        # construct response
+        response = []
+        response.append({
+            "user_id": user.user_id,                
+            "user_first_name": user.first_name,
+            "user_last_name": user.last_name,
+            "user_username": user.username,
+            "user_email": user.email,
+            "user_date_created": user.date_created
+        })        
+
+        return jsonify(response), 200, headers
+
+    # user with user_id does not exist in the database
+    response = {
+        "message": f"User with user_id {user_id} does not exist in the database. Unable to get user information.",
+        "success": False
+    }
+
+    return jsonify(response), 400, headers
 
 
 @users_bp.route("/delete_user", methods=["DELETE"])
 # @login_required
 def delete_user():
-    """Delete user from database.
-    
-    Precondition: user_id exists in the 'user' table
-    """
+    """Delete user from database."""
 
     # get arguments
     user_id = request.args.get("user_id")
 
-    # delete user
-    user = User.query.get(user_id)
-    db.session.delete(user)    
-    db.session.commit()    
-    return make_response(f"{user} successfully deleted.", 200)
+    # prepare headers for response
+    headers = {"Content-Type": "application/json"}
+
+    # delete user if user_id is found
+    if user_id:
+        user = User.query.get(user_id)
+        db.session.delete(user)    
+        db.session.commit()    
+
+        response = {
+            "message": f"{user} successfully deleted.",
+            "success": True
+        }
+        
+        return jsonify(response), 200, headers
+    
+    # user with user_id does not exist in the database
+    response = {
+        "message": f"User with user_id {user_id} does not exist in the database. Unable to delete.",
+        "success": False
+    }
+
+    return jsonify(response), 400, headers
 
 
 @users_bp.route("/update_user", methods=["PUT"])
 # @login_required
 def update_user():
-    """Update details of a user.
-    
-    Precondition(s):
-    - user_id is not null
-    """
+    """Update details of a user."""
 
     # get arguments
     user_id = request.args.get("user_id")
@@ -117,49 +141,60 @@ def update_user():
     # get desired user
     user  = User.query.get(user_id)
 
-    # get details about current user in database
-    curr_username = user.username
-    curr_email = user.email
-
     # prepare headers for response
     headers = {"Content-Type": "application/json"}
 
-    # make sure username don't already exist in database (excluding user's current username)
-    if username != curr_username and User.query.filter_by(username=username).count() > 0:
-        # return make_response(f"Unable to update user. Username \"{username}\" is already taken.")
-        response = {
-            "message": f"Unable to update user. Username \"{username}\" is already taken.",
-            "success": False
-        }
-        return jsonify(response), 409
+    # update user information is user_id is found
+    if user:
 
-    # make sure email don't already exist in database (excluding user's current email)
-    if email != curr_email and User.query.filter_by(email=email).count() > 0:
-        # return make_response(f"Unable to update email. \"{email}\" is already associated with another account.")
-        response = {
-            "message": f"Unable to update email. \"{email}\" is already associated with another account.",
-            "success": False
-        }
-        return jsonify(response), 409
+        # get details about current user in database
+        curr_username = user.username
+        curr_email = user.email        
 
-    # update user details
-    if len(first_name) != 0 and first_name is not None:
-        user.first_name = first_name
-    if len(last_name) != 0 and last_name is not None:
-        user.last_name = last_name    
-    if len(username) != 0 and username is not None:
-        user.username = username    
-    if len(email) != 0 and email is not None: # add email format verification here or in client-side code?
-        user.email = email
-    if len(password) != 0 and password is not None:
-        user.password = generate_password_hash(password)
+        # make sure username don't already exist in database (excluding user's current username)
+        if username != curr_username and User.query.filter_by(username=username).count() > 0:
+            # return make_response(f"Unable to update user. Username \"{username}\" is already taken.")
+            response = {
+                "message": f"Unable to update user. Username \"{username}\" is already taken.",
+                "success": False
+            }
+            return jsonify(response), 409, headers
+
+        # make sure email don't already exist in database (excluding user's current email)
+        if email != curr_email and User.query.filter_by(email=email).count() > 0:
+            # return make_response(f"Unable to update email. \"{email}\" is already associated with another account.")
+            response = {
+                "message": f"Unable to update email. \"{email}\" is already associated with another account.",
+                "success": False
+            }
+            return jsonify(response), 409, headers
+
+        # update user details
+        if len(first_name) != 0 and first_name is not None:
+            user.first_name = first_name
+        if len(last_name) != 0 and last_name is not None:
+            user.last_name = last_name    
+        if len(username) != 0 and username is not None:
+            user.username = username    
+        if len(email) != 0 and email is not None: # add email format verification here or in client-side code?
+            user.email = email
+        if len(password) != 0 and password is not None:
+            user.password = generate_password_hash(password)
+        
+        # update database entry
+        db.session.commit()
+        
+        # return make_response(f"{user} successfully updated.", 200)
+        response = {
+            "message": f"{user} successfully updated.",
+            "success": True
+        }
+        return jsonify(response), 200, headers
     
-    # update database entry
-    db.session.commit()
-    
-    # return make_response(f"{user} successfully updated.", 200)
+    # user with user_id does not exist in the database
     response = {
-        "message": f"{user} successfully updated.",
-        "success": True
+        "message": f"User with user_id {user_id} does not exist in the database. Unable to update.",
+        "success": False
     }
-    return jsonify(response), 200, headers
+
+    return jsonify(response), 400, headers
